feat: add new brand icons (Scira, V0, X, YouTube) and refactor icon utilities
