+++ src/components/icons/brands/grok.tsx
@@ -1,0 +1,15 @@
+import type { IconProps } from '../utils'
+import { getSvgSizeProps } from '../utils'
+
+export const GrokIcon = ({ height, size, width, ...props }: IconProps) => (
+  <svg
+    aria-hidden
+    fill='currentColor'
+    viewBox='0 0 24 24'
+    xmlns='http://www.w3.org/2000/svg'
+    {...getSvgSizeProps({ height, size, width })}
+    {...props}
+  >
+    <path d='m9.269 14.855 7.979-5.923c.39-.29.95-.177 1.136.274.981 2.379.543 5.237-1.409 7.2s-4.667 2.393-7.15 1.413l-2.71 1.262c3.888 2.674 8.61 2.013 11.562-.957 2.34-2.354 3.066-5.563 2.388-8.457l.006.007c-.983-4.251.242-5.95 2.75-9.425Q23.912.126 24 0l-3.302 3.32v-.01L9.267 14.857M7.622 16.295c-2.79-2.682-2.31-6.832.072-9.225C9.455 5.3 12.341 4.576 14.86 5.64l2.705-1.256a7.8 7.8 0 0 0-1.829-1.003 8.95 8.95 0 0 0-9.752 1.973C3.451 7.9 2.654 11.817 4.022 15.16c1.022 2.498-.653 4.265-2.34 6.049C1.082 21.84.482 22.473 0 23.143l7.62-6.846' />
+  </svg>
+)
