+++ src/components/icons/brands/index.ts
@@ -1,0 +1,12 @@
+export * from './claude'
+export * from './cursor'
+export * from './github'
+export * from './google'
+export * from './grok'
+export * from './linkedin'
+export * from './markdown'
+export * from './openai'
+export * from './scira'
+export * from './v0'
+export * from './x'
+export * from './youtube'
